web/requirements.txt
Django==1.10.7
django-common-helpers==0.8.0
django-compressor==2.1
django-cors-headers==2.2.0
django-extensions==1.7.4
django-filter==1.1.0
django-ical==1.4
django-localflavor==1.3
django-nested-admin==3.0.21
django-registration==2.1.2
djangorestframework==3.8.2
Fabric==1.12.0
# MySQL-python==1.2.3
mysqlclient
Pillow==4.1.1
pycryptodome
simplejson==3.8.2
sorl-thumbnail==12.3
django-debug-toolbar==1.5